apps/api/pyproject.toml
[project]
name = "platform-api"
version = "0.1.0"
description = "AIpacken AI Platform — FastAPI backend + Arq worker + builder service"
readme = "../../README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.4",
  "uvicorn[standard]==0.32.0",
  "pydantic==2.9.2",
  "pydantic-settings==2.6.1",
  "sqlalchemy==2.0.36",
  "alembic==1.13.3",
  "psycopg[binary]==3.2.3",
  "arq==0.26.1",
  "redis==5.1.1",
  "boto3==1.35.0",
  "python-multipart==0.0.17",
  "itsdangerous==2.2.0",
  "structlog==24.4.0",
  "httpx==0.27.2",
  "jinja2==3.1.4",
  "docker==7.1.0",
  "mlflow-skinny==2.17.0",
  "pandas==2.2.3",
  "pyarrow==17.0.0",
  "ydata-profiling==4.11.0",
  "scikit-learn==1.5.2",
  "shap==0.46.0",
  "fairlearn==0.11.0",
  "watchfiles==0.24.0",
  "passlib[bcrypt]==1.7.4",
  "python-jose[cryptography]==3.3.0",
  "aiosqlite==0.20.0",
  "httpx-sse==0.4.0",
  "uuid6==2024.7.10",
  "email-validator==2.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.3",
  "pytest-asyncio==0.24.0",
  "httpx==0.27.2",
  "ruff==0.7.2",
  "mypy==1.13.0",
  "types-requests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aipacken"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "S", "C4", "PL", "RUF"]
ignore = ["S101", "PLR0913"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
